requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
crispy-bootstrap5==0.6
Django==3.2.11
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==21.2
django-bootstrap5==21.3
django-crispy-forms==1.13.0
django-dynamic-preferences==1.11.0
django-macros==0.4.0
django-simple-history==3.0.0
django-sslserver==0.22
django-user-activity-log2==0.0.24
idna==3.3
numpy==1.22.0
persisting-theory==0.2.1
Pillow==9.0.0
pytz==2021.3
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.7
